Use size_t for rcAlloc and dtAlloc
This was already partly done by @cexikitin in #65, I'm just finishing it
off in their absence!

Closes #65

+++ DetourTileCache/Include/DetourTileCacheBuilder.h
@@ -84,7 +84,7 @@ struct dtTileCacheAlloc
 	{
 	}
 	
-	virtual void* alloc(const int size)
+	virtual void* alloc(const size_t size)
 	{
 		return dtAlloc(size, DT_ALLOC_TEMP);
 	}
